requirements.txt
torch
torchvision
pillow
matplotlib
tensorboard